requirements.txt
prometheus_client
pyserial